# lab4-consumingRESTServices2/src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers: 
  - url: 'https://easy-trip-350018.ew.r.appspot.com/api/v1' 
info:
  description: EASY TRIP is an API REST Project made with Java, the purpose of this api is to facilitate the creation and use of flights and passengers.
  version: 1.0.0
  title: EASY-TRIP
  contact:
    email: miggonort1@alum.us.es
  license:
    name: Github Project
    url: 'https://github.com/miggonort1/EASY-TRIP'
...
paths:
  /passengers:
    get:
      tags:
        - Passengers
      summary: Get all the passengers
      description: Get all the existing passengers
      operationId: getAllPassengers
      parameters:
        - name: order
          in: query
          description: Order results by name, surname or age
          required: false
          schema:
            type: string
            enum:
              - name
              - '-name'
              - surname
              - '-surname'
              - age
              - '-age'
        - name: q
          in: query
          description: Query to filter
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description:  Maximum number of results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: Offset of the results
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - name
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    surname:
                      type: string
                    age:
                      type: string
                  example:
                    id: p1
                    name: Roberto
                    surname: García Martínez
                    age: '23'
        '400':
          description: Bad request
    post:
      tags:
        - Passengers
      summary: Create a passenger
      description: Create a new passenger
      operationId: addPassenger
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
                - surname
                - age
              properties:
                id:
                  type: string
                name:
                  type: string
                surname:
                  type: string
                age:
                  type: string
              example:
                id: p40
                name: Alejandro
                surname: Sayago Rico
                age: '33'
        description: Passenger to create
        required: true
      responses:
        '201':
          description: Passenger created
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  surname:
                    type: string
                  age:
                    type: string
                example:
                  id: p40
                  name: Alejandro
                  surname: Sayago Rico
                  age: '33'
        '400':
          description: Bad request
    put:
      tags:
        - Passengers
      summary: Update a passenger
      description: Update an existing passenger
      operationId: updatePassenger
      requestBody:
        content:
          application/json:
            schema:
              required:
                - id
                - name
                - surname
                - age
              properties:
                id:
                  type: string
                name:
                  type: string
                surname:
                  type: string
                age:
                  type: string
              example:
                id: p5
                name: Carlos
                surname: Conde Sánchez
                songs: '23'
        description: Passenger to update
      responses:
        '204':
          description: Passenger updated
        '400':
          description: Bad request
        '404':
          description: Passenger not found
  '/passengers/{passengerId}':
    get:
      tags:
        - Passengers
      summary: Get a passenger
      description: Get an existing passenger
      operationId: getPassenger
      parameters:
        - name: passengerId
          in: path
          description: Id of the passenger
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Passenger found
          content:
            application/json:
              schema:
                required:
                  - name
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  surname:
                    type: string
                  age:
                    type: array
                example:
                  id: p40
                  name: Carlos
                  surname: Conde Sánchez
                  age: '33'
        '404':
          description: Passenger not found
    delete:
      tags:
        - Passengers
      summary: Delete a passenger
      description: Delete an existing passenger
      operationId: deletePassenger
      parameters:
        - name: passengerId
          in: path
          description: Id of the passenger
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Passenger deleted
        '404':
          description: Passenger not found
...